requirements.txt
Django==4.2.13
djangorestframework==3.15.1
joblib==1.3.2
joblib==1.4.0
numpy==2.0.0
pandas==2.2.2
scikit_learn==1.4.1.post1
